Move Files To Landing Page Folder For Better structure
