+++ src/components/landing-page/landing-top-blog/TopBlogCard.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const TopBlogCard = () => {
+  return (
+    <div>TopBlogCard</div>
+  )
+}
+
+export default TopBlogCard
