+++ src/components/landing-page/landing-top-blog/TopBlogsDiv.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const TopBlogsDiv = () => {
+  return (
+    <div>TopBlogsDiv</div>
+  )
+}
+
+export default TopBlogsDiv
